queue: move com.ning.billing.QueueObjectMapper to com.ning.billing.queue.QueueObjectMapper
This fixes the following bug in OSGI plugins:

   Caused by: java.lang.LinkageError: loader constraint violation: loader
   (instance of org/apache/felix/framework/BundleWiringImpl$BundleClassLoaderJava5)
   previously initiated loading for a different type with name "com/fasterxml/jackson/databind/ObjectMapper"

In Kill Bill, com.ning.billing is exported to OSGI plugins. During the
initialization process in beatrix, the static ObjectMapper in QueueObjectMapper
was created with the Webapp classloader. When OSGI plugins would then
create their own queues, they'd share the same QueueObjectMapper (due to
the unfortunate poorly chosen package name), while the rest of the
queue code would be created by the BundleClassLoader.

The mismatch when loading the ObjectMapper from the plugin would then
cause the linkage error.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>