See https://github.com/killbill/killbill-commons/issues/6
After re-reading the code, we realized we were missing a synchronized block for the case where we run in sticky mode and we use batchUpdate. Instead of adding that synchonized block, we decided to simplify the code and synchronize the method getReadyEntries() because there were already 2 synchronized block in the method, and therefore at this point this is simpler and pretty much equivalent in terms of performance to synchronized the whole method.

Hoepfully this addresses the issue, but since we have a hard time to repro the issue, we don't mark the bug as fixed yet.