queue: validate auto_increment column at startup
The auto_increment column in the current tables (e.g. bus_events)
could be reset accidentally (in case of a TRUNCATE for instance).
This would cause errors when moving further entries to history.

This patch ensures the column is reset at startup if needed,
by setting it to the max record_id found in the history table.

See https://github.com/killbill/killbill-commons/issues/6.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>