Commit 31db511cb037be22116e217d7446d843930b4e1a removed the auto_increment for the bus_events_history table; that lead to some issues because mysql only keeps an in-memory version so any restart will lose the count and prevent entries to be moved to history.
Commit 23e792c59a45511eeff97faae8ee4fcdac60d960 was an attempt to fix the issue by resetting the auto_increment of the bus_events table based on where we are in
bus_events history table.

This patch reverts to all behavior where both bus_events and bus_events _history have an auto_increment (which means record_id for a given entry in the bus_events
may not match the record_id in the bus_events_history table), but at least this will not lead to band-aid attempting to fix mysql behavior.