embeddeddb: fork testing-postgresql-server
Because of the Guice wiring in the Kill Bill tests, we need to access
the database port before the database is started.

Classes from testing-postgresql-server had to be copied over due to their
limited visibility.

This also makes sure the bridge works on Java 6 (although it is possible
that some of the airlift dependencies are Java 7 only - to be tested).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>