queue: remove trailing ; in batch query
For some reason, this is sometimes making the batch insert fail.
I was able to reproduce it on one laptop, but not another one, under a similar setup.

According to git bisect, the culprit seems to be bbeb3af036345d630d5f29501697ed8da9ceac37,
although I'm unclear how it's related.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>