queue: add missing setters in ModelDao
When mapping results, LowerToCamelBeanMapper will first try
to use the setter. If one isn't found, getDeclaredField is called
instead -- this is quite costly because when NoSuchFieldException
is thrown (often, due to the inheritance NotificationEventModelDao < BusEventModelDao),
fillInStackTrace is called.

Adding these setters lead to significant speedups in some of our load tests.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>