jdbi: enhance caching of StringTemplates when a couple of attributes are defined
This is worth the complexity, as such queries are heavily used for Queues.

To make the caching work though, we have to rework how IN clauses are generated
(basically, pass a collection of ready-to-be-bound attributes to StringTemplate instead
of having the template do the transformation itself). The easiest is just to use @BindIn
instead of our custom collection binders.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>