queue: Introduce 2 new configs for inflightQ to indicate min and max amount of entries to extract.
The 'maxInFlightEntries' setting indicates the max # of records we will pull from the DB at a time.
The 'minInFlightEntries' setting indicates the min # of records we would like to use to fetch entries from the DB (to avoid too many too small queries).
However, if the rate of incoming request is slow, we will proceed with whatever we have after waiting a maximum of 100 mSec (hardcoded).