queue: revisit processing_available_date usage
processing_available_date used in the readyWhereClause was useful
to detect stuck entries. This is now redundant since we introduced
the reaper.

processing_available_date is still useful for the reaper when
a queue is late: if an entry is IN_PROCESSING (i.e. a node has claimed it),
we don't want to reap it unless it is truly stuck.

Note also that processing_available_date is updated before moving entries into
the history table, so the column is useful to determine when the entry was processed
(processing_available_date - created_date is the total processing time for the bus).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>