queue: revisit shutdown sequence
Make timeout configurable to wait for executors threads to complete.

Some handlers can take a long time (e.g. Analytics plugin handlers) and
the default hardcoded 5s value isn't always enough in practice. This
can result in leaving IN_PROCESSING entries on disk if the shutdown
sequence doesn't wait for these threads to finish.

As part of this, the shutdown sequence was revisited too. With the previous
implementation, even if the shutdown sequence waits for these threads to finish,
IN_PROCESSING entries could still be left on disk if the completion threads are
stopped before all jobs in the dispatcher queue are done.

I've also bumped the default timeout from 5s to 15s (5s might be a bit aggressive,
even for core Kill Bill handlers).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>